fix(seo): define `/blog` metadata on `page.tsx` instead of `layout.tsx`
Prevent pass metadata to subroutes of `/blog`
