+++ src/app/blog/layout.tsx
@@ -1,20 +1,3 @@
-import type { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: 'Blog',
-  description:
-    'My personal posts and thoughts about all kind of stuff that I like. Main philosophic and technical text.',
-  keywords: [
-    'blog',
-    'thoughts',
-    'texts',
-    'technical',
-    'tutorials',
-    'posts',
-    'phisolophy'
-  ]
-}
-
 export default function BlogLayout({
   children
 }: {
